- Used soft delete pattern for data - Implemented GetNewTracks and GetNewAlbums - Added TracksArtist table, and now track can have multiple authors - Adapted to the new database - Implemented NewTracksRequest, NewAlbumsRequest, ArtistTracksResponse, ArtistResponse, ArtistAlbumsResponse, AlbumRequest - Updated db diagram - Made fixes and improvement

+++ HomeLib/HomeLib.Services/Services/TrackService.cs
@@ -15,9 +15,9 @@ public class TrackService(ITrackRepository trackRepository) : ITrackService
 
     public async Task<Track> GetTrackById(Guid id)
     {
-        if(id == Guid.Empty) throw new BadRequestException("Invalid track id");
+        if (id == Guid.Empty) throw new BadRequestException("Invalid track id");
         var track = await trackRepository.GetTrackByIdAsync(id);
-        
+
         return track ?? throw new NotFoundException($"Track with {id} not found");
     }
 
@@ -29,30 +29,59 @@ public class TrackService(ITrackRepository trackRepository) : ITrackService
             Name = name,
             Duration = duration,
             AlbumId = albumId,
-            ArtistId = artistId,
+            CreatedAt =  DateTime.UtcNow,
+            UpdatedAt = DateTime.UtcNow,
         };
 
-        await trackRepository.AddTrackAsync(track);
+        var trackArtist = new TracksArtists()
+        {
+            ArtistId = artistId,
+            TrackId = track.Id,
+            UpdatedAt = track.UpdatedAt,
+        };
+        await trackRepository.AddTrackAsync(track, trackArtist);
     }
 
-    public async Task DeleteTrack(Guid id)
+    public async Task HardDeleteTrack(Guid id)
     {
-        if(id == Guid.Empty) throw new BadRequestException("Invalid track id");
+        if (id == Guid.Empty) throw new BadRequestException("Invalid track id");
         var track = await trackRepository.GetTrackByIdAsync(id);
         if (track == null) throw new NotFoundException($"Track with {id} not found");
-        
+        var trackInTrackArtist = await trackRepository.DeleteTrackArtistAsync(id);
+        if (trackInTrackArtist == null) throw new NotFoundException($"Track with {id} not found");
+
+        trackInTrackArtist.IsDeleted = true;
+
         await trackRepository.DeleteTrackAsync(id);
+    }
+
+    public async Task SoftDeleteTrack(Guid id)
+    {
+        if (id == Guid.Empty) throw new BadRequestException("Invalid track id");
+        var track = await trackRepository.GetTrackByIdAsync(id);
+        if (track == null) throw new NotFoundException($"Track with {id} not found");
+
+        track.IsDeleted = true;
+        track.UpdatedAt = DateTime.UtcNow;
+
+        await trackRepository.UpdateTrackAsync(track);
     }
 
     public async Task UpdateTrack(Guid id, string name, int duration)
     {
-        if(id == Guid.Empty) throw new BadRequestException("Invalid track id");
+        if (id == Guid.Empty) throw new BadRequestException("Invalid track id");
         var track = await trackRepository.GetTrackByIdAsync(id);
         if (track == null) throw new NotFoundException($"Track with {id} not found");
 
         track.Name = name;
         track.Duration = duration;
-        
+        track.UpdatedAt = DateTime.UtcNow;
+
         await trackRepository.UpdateTrackAsync(track);
+    }
+
+    public async Task<List<Track>> GetNewTracks(int count)
+    {
+        return await trackRepository.GetNewTracks(count);
     }
 }
